feat: refactor TreasurePage to use WalletPage component and create UserWalletPage for dynamic wallet address rendering
